Remove isPrivateMode from BrowserChecker
