+++ universal-login-react/src/index.ts
@@ -41,6 +41,5 @@ export {WaitingForTransaction, WaitingForTransactionProps, WaitingForDeployment,
 export {WaitingForOnRampProvider, WaitingForOnRampProviderProps} from './ui/TopUp/Fiat/WaitingForOnRampProvider';
 export {ThemeProvider, ThemeContext} from './ui/themes/Theme';
 export {calculateTransactionFee} from './core/utils/calculateTransactionFee';
-export {isPrivateMode} from './ui/utils/isPrivateMode';
-export {isLocalStorageBlocked} from './ui/utils/isLocalStorageBlocked';
 export {CompanyLogo} from './ui/commons/CompanyLogo';
+export {BrowserChecker} from './app/BrowserChecker';
