--- universal-login-react/src/ui/utils/isLocalStorageBlocked.ts
@@ -1,17 +1,0 @@
-export function isLocalStorageBlocked() {
-  try {
-    if (!('localStorage' in window)) {
-      return true;
-    }
-    if (window.localStorage === undefined) {
-      return true;
-    }
-    localStorage.setItem('localStorageTest', 'foo');
-    localStorage.removeItem('localStorageTest');
-    window.localStorage.setItem('localStorageTest', 'foo');
-    window.localStorage.removeItem('localStorageTest');
-    return false;
-  } catch {
-    return true;
-  }
-}
