+++ universal-login-web3/test/services/setupStrategies.test.ts
@@ -4,7 +4,6 @@ import {setupStrategies} from '../../src/services/setupStrategies';
 import {Web3ProviderFactory} from '../../src/models/Web3ProviderFactory';
 import UniLoginLogo from '../../src/ui/assets/U.svg';
 import MetamaskLogo from '../../src/ui/assets/MetaMaskLogoTitle.svg';
-import {BrowserChecker} from '../../src/services/BrowserChecker';
 
 describe('UNIT: setupStrategies', () => {
   const mockProvider = {send: (_: any, cb: any) => cb(null, {result: '1'})};
@@ -15,7 +14,7 @@ describe('UNIT: setupStrategies', () => {
       logo: 'https://kickback.events/favicon.ico',
       type: 'laptop',
     };
-    const result = setupStrategies(mockProvider, ['UniLogin'], {applicationInfo, storageService: new MemoryStorageService(), browserChecker: {isLocalStorageBlocked: () => false} as BrowserChecker});
+    const result = setupStrategies(mockProvider, ['UniLogin'], {applicationInfo, storageService: new MemoryStorageService(), browserChecker: {isLocalStorageBlocked: () => false}});
     expect(result[0].name).to.eq('UniLogin');
     expect(result[0].icon).to.eq(UniLoginLogo);
     expect(result[0].create).to.not.be.null;
